pyproject.toml
[project]
name = "prodik"
version = "0.1.0"
description = "Great & powerful application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.5",
    "dishka>=1.7.2",
    "fastapi>=0.128.0",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.46",
    "structlog>=25.5.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "ruff",
    "datamodel-code-generator>=0.54.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = [".venv", "src/prodik/infrastructure/db/migrations"]

[tool.ruff.lint]
extend-select = ["ALL"]
ignore = [
    "ERA001",
    "T201",

    "PT019",
    "ANN401",
    "COM812",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D203",
    "D212",
    "EM101",
    "EM102",
    "E731",
    "TRY003",
    "RUF001",
    "RUF002",
    "RUF003",
    "TC003",
    "ISC001",
    "D205",
    "S101",
    "PGH003",
    "RUF100",
    "C408",
    "FIX002",
    "TD003",
    "TD002",
    "TD004",
    "TC002"
]
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"


[tool.mypy]
strict = true
explicit_package_bases = true
exclude = "build"
warn_unreachable = true
show_column_numbers = true
show_error_context = true
check_untyped_defs = true
ignore_missing_imports = false
warn_no_return = true
warn_return_any = true
mypy_path = "src"
plugins = [
    "pydantic.mypy",
]
files = ["src/prodik"]


[project.scripts]
prodik = "prodik.bootstrap.cli:main"
